Use Tailwind to extract actual CSS rules
Based on the code in Devzstudio/tailwind_to_css#47

Caveat: By configuring a maximally broad safeList, all CSS classes from Tailwind will be included. This can make compilation and deployment slow, and there will be a large CSS file to download when loading the website. This is unavoidable for this new technique to work.

Also:
- Make the JSON output pretty.
- Delay processing by 1 second when typing the input, otherwise the text box is slow to show what you're typing.