Switch to server-side API to generate classes
This will dynamically generate the Tailwind result CSS, ensuring the class to CSS rule conversion is complete, including arbitrary values in square brackets.

This also reverts the safeList changes. We will no longer need to spend a lot of time deploying, or downloading a giant list of classes. Yay, Performance Restored!!!

This is based on the new code in Devzstudio/tailwind_to_css#47.

Also: Trim the CSS result when copying.